move all filtering logic to the provider
